requirements.txt
django
gunicorn
pillow
django-cleanup
beautifulsoup4
whitenoise
django-environ
requests
sendgrid
